pyproject.toml
[tool.black]
line-length = 79

[tool.djlint]
indent = 2
profile = "jinja"
max_line_length = 79
ignore = "H021"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.poetry]
name = "autobidsportal"
version = "2.1.0"
description = ""
authors = ["Tristan Kuehn <tkuehn@uwo.ca>"]
readme = "README.md"
packages = [
    { include = "autobidsportal" },
    { include = "autobidsportal_migrations" }
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4"
backports-zoneinfo = { version = "^0.2.1", python = "<3.9" }
datalad = "^0.18.1"
Flask = "~2.2.2"
Flask-Excel = "^0.0.7"
Flask-Login = "^0.6.2"
Flask-Migrate = "^4.0.4"
Flask-SQLAlchemy = "^3.0.3"
SQLAlchemy = "<2.0"
Flask-WTF = "~1.0.1"
Flask-Mail = "^0.9.1"
rq = "^1.12.0"
defusedxml = "^0.7.1"
email-validator = "^1.3.1"
Werkzeug = "^2.2.2"
WTForms = "^2.3.3"
pyuwsgi = { version = "^2.0.21", optional = true }
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pyright = "^1.1"
djlint = "^1.19.13"
pytest = "^7.2.1"
ruff = "^0.0.265"

[tool.ruff]
select = ["ALL"]
ignore = ["ANN", "TCH", "T20", "S603", "S607", "B905"]
unfixable = ["F841"]

[tool.poetry.extras]
deploy = ["pyuwsgi"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
